Added even more functionality and tests

+++ main.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"bufio"
 	"bytes"
 	"errors"
 	"flag"
@@ -13,6 +14,7 @@ import (
 	"syscall"
 
 	"github.com/GregorioDiStefano/gcloud-fuse/simplecrypto"
+	"github.com/ryanuber/go-glob"
 	"golang.org/x/crypto/ssh/terminal"
 	"golang.org/x/net/context"
 	"golang.org/x/oauth2/google"
@@ -20,6 +22,7 @@ import (
 )
 
 func init() {
+	flag.Bool("i", false, "interactive mode")
 	flag.Bool("list", false, "list folders/files")
 	flag.String("delete", "", "delete object")
 	flag.String("download", "", "file to download to local disk")
@@ -30,7 +33,6 @@ func init() {
 
 func main() {
 	flag.Parse()
-
 	userData := parseConfig()
 
 	fmt.Print("Password: ")
@@ -67,7 +69,15 @@ func main() {
 		os.Exit(1)
 	}
 
-	parseCmdLine(bs, *cryptoKeys)
+	if flag.Lookup("i").Value.String() == "true" {
+		for {
+			reader := bufio.NewReader(os.Stdin)
+			fmt.Print("> ")
+			reader.ReadString('\n')
+		}
+	} else {
+		parseCmdLine(bs, *cryptoKeys)
+	}
 	os.Exit(0)
 }
 
@@ -76,18 +86,18 @@ func verifyPassword(bs *bucketService, cryptoKeys simplecrypto.Keys) error {
 	testdata, err := simplecrypto.EncryptText(testString, cryptoKeys.EncryptionKey)
 
 	if err != nil {
-		return errors.New("Unable to encrypt test string: " + err.Error())
+		return errors.New("unable to encrypt test string: " + err.Error())
 	}
 
 	testfile, err := bs.downloadFromBucket("keycheck")
 	defer os.Remove(testfile)
 
 	if err != nil {
-		return errors.New(fmt.Sprintf("Failed to find a \"keycheck\" file, if this is a new bucket, create a file called, \"keycheck\" containing: %s", testdata))
+		return errors.New(fmt.Sprintf("failed to find a \"keycheck\" file, if this is a new bucket, create a file called, \"keycheck\" containing: %s", testdata))
 	} else {
 		testfileBytes, _ := ioutil.ReadFile(testfile)
 		if plainText, err := simplecrypto.DecryptText(string(testfileBytes), cryptoKeys.EncryptionKey); err != nil || plainText != testString {
-			return errors.New("Failed to verify bucket is using specified password: " + err.Error())
+			return errors.New("failed to verify bucket is using specified password: " + err.Error())
 		}
 	}
 	return nil
@@ -101,18 +111,7 @@ func parseCmdLine(bs *bucketService, cryptoKeys simplecrypto.Keys) {
 		returnedError = doDeleteObject(bs, cryptoKeys, flag.Lookup("delete").Value.String())
 	case flag.Lookup("upload").Value.String() != "":
 		path := flag.Lookup("upload").Value.String()
-
-		if isDir(path) {
-			returnedError = filepath.Walk(path, func(path string, info os.FileInfo, err error) error {
-				if !isDir(path) {
-					return doUpload(bs, cryptoKeys, path, flag.Lookup("dir").Value.String())
-				}
-				return nil
-			})
-		} else {
-			returnedError = doUpload(bs, cryptoKeys, path, flag.Lookup("dir").Value.String())
-		}
-
+		returnedError = processUpload(bs, path, cryptoKeys)
 	case flag.Lookup("download").Value.String() != "":
 		returnedError = doDownload(bs, cryptoKeys, flag.Lookup("download").Value.String())
 	case flag.Lookup("list").Value.String() == "true":
@@ -123,6 +122,40 @@ func parseCmdLine(bs *bucketService, cryptoKeys simplecrypto.Keys) {
 		fmt.Println("Action returned error: " + returnedError.Error())
 		os.Exit(1)
 	}
+}
+
+func processUpload(bs *bucketService, path string, cryptoKeys simplecrypto.Keys) error {
+	globMatch, err := filepath.Glob(path)
+	var errorUploading bool
+	fmt.Println(globMatch, err)
+	if err != nil {
+		return err
+	}
+
+	for _, path := range globMatch {
+		fmt.Println("Uploading: " + path)
+		if isDir(path) {
+			err = filepath.Walk(path, func(path string, info os.FileInfo, err error) error {
+				if !isDir(path) {
+					fmt.Println("Uploading: " + path)
+					return doUpload(bs, cryptoKeys, path, flag.Lookup("dir").Value.String())
+				}
+				return nil
+			})
+		} else {
+			err = doUpload(bs, cryptoKeys, path, flag.Lookup("dir").Value.String())
+		}
+
+		if err != nil {
+			errorUploading = true
+			fmt.Println(fmt.Sprintf("failed with %s when uploading: %s", err.Error(), path))
+		}
+	}
+
+	if errorUploading {
+		return errors.New("at least one file failed to upload")
+	}
+	return nil
 }
 
 func doUpload(bs *bucketService, keys simplecrypto.Keys, filePath, remoteDirectory string) error {
@@ -149,8 +182,13 @@ func doUpload(bs *bucketService, keys simplecrypto.Keys, filePath, remoteDirecto
 		fmt.Println(err)
 	}
 
-	hmac := simplecrypto.CalculateHMAC(keys.HMACKey, iv, encryptedFile, false)
-	addHMACToFile(encryptedFile, hmac)
+	hmac, err := simplecrypto.CalculateHMAC(keys.HMACKey, iv, encryptedFile, false)
+
+	if err != nil {
+		return err
+	}
+
+	simplecrypto.AddHMACToFile(encryptedFile, hmac)
 	bs.uploadToBucket(encryptedFile, encryptedPath)
 	return nil
 }
@@ -158,7 +196,7 @@ func doUpload(bs *bucketService, keys simplecrypto.Keys, filePath, remoteDirecto
 func printList(bs *bucketService, key []byte) {
 	objects, err := bs.getObjects()
 	if err != nil {
-		panic("Failed getting objects:" + err.Error())
+		panic("failed getting objects:" + err.Error())
 	}
 
 	encToDecPaths := getEncryptedToDecryptedMap(objects, key)
@@ -174,41 +212,52 @@ func doDownload(bs *bucketService, keys simplecrypto.Keys, filename string) erro
 	objects, err := bs.getObjects()
 
 	if err != nil {
-		return errors.New("Failed getting objects: " + err.Error())
+		return errors.New("failed getting objects: " + err.Error())
 	}
 
 	encToDecPaths := getEncryptedToDecryptedMap(objects, keys.EncryptionKey)
 
-	encryptedFilepath := encToDecPaths[filename]
-	decryptedFilePath := decryptFilePath(encToDecPaths[filename], keys.EncryptionKey)
-	downloadedEncryptedFile, _ := bs.downloadFromBucket(encryptedFilepath)
-	defer os.Remove(downloadedEncryptedFile)
+	for k, _ := range encToDecPaths {
+		//TODO: should be decToEncPaths
+		fmt.Println(filename, k)
+		if glob.Glob(filename, k) {
+			fmt.Println("downloading: " + k)
+			encryptedFilepath := encToDecPaths[k]
+			decryptedFilePath := decryptFilePath(encToDecPaths[k], keys.EncryptionKey)
+			downloadedEncryptedFile, _ := bs.downloadFromBucket(encryptedFilepath)
+			defer os.Remove(downloadedEncryptedFile)
 
-	iv, err := simplecrypto.GetIVFromEncryptedFile(downloadedEncryptedFile)
+			iv, err := simplecrypto.GetIVFromEncryptedFile(downloadedEncryptedFile)
 
-	if err != nil {
-		panic(err)
-	}
+			if err != nil {
+				panic(err)
+			}
 
-	actualHMAC := simplecrypto.CalculateHMAC(keys.HMACKey, iv, downloadedEncryptedFile, true)
+			actualHMAC, err := simplecrypto.CalculateHMAC(keys.HMACKey, iv, downloadedEncryptedFile, true)
 
-	if expectedHMAC, err := getHMACFromFile(downloadedEncryptedFile); err == nil {
-		if !bytes.Equal(actualHMAC, expectedHMAC) {
-			panic("File has been tampered with!")
+			if err != nil {
+				return err
+			}
+
+			if expectedHMAC, err := simplecrypto.GetHMACFromFile(downloadedEncryptedFile); err == nil {
+				if !bytes.Equal(actualHMAC, expectedHMAC) {
+					panic("file has been tampered with!")
+				}
+			} else {
+				panic(err.Error())
+			}
+
+			downloadedPlaintextFile, _ := simplecrypto.DecryptFile(downloadedEncryptedFile, keys.EncryptionKey)
+			defer os.Remove(downloadedPlaintextFile)
+
+			_, tempDownloadFilename := path.Split(downloadedPlaintextFile)
+			_, actualFilename := path.Split(decryptedFilePath)
+
+			os.Rename(tempDownloadFilename, actualFilename)
+			simplecrypto.TruncateHMACSignature(actualFilename)
+			fmt.Println("downloaded: " + actualFilename)
 		}
-	} else {
-		panic(err.Error())
 	}
-
-	downloadedPlaintextFile, _ := simplecrypto.DecryptFile(downloadedEncryptedFile, keys.EncryptionKey)
-	defer os.Remove(downloadedPlaintextFile)
-
-	_, tempDownloadFilename := path.Split(downloadedPlaintextFile)
-	_, actualFilename := path.Split(decryptedFilePath)
-
-	os.Rename(tempDownloadFilename, actualFilename)
-	truncateHMACSignature(actualFilename)
-	fmt.Println("Downloaded: " + actualFilename)
 	return nil
 }
 
@@ -216,7 +265,7 @@ func doDeleteObject(bs *bucketService, keys simplecrypto.Keys, filename string) 
 	objects, err := bs.getObjects()
 
 	if err != nil {
-		return errors.New("Failed getting objects: " + err.Error())
+		return errors.New("failed getting objects: " + err.Error())
 	}
 
 	encToDecPaths := getEncryptedToDecryptedMap(objects, keys.EncryptionKey)
@@ -224,7 +273,7 @@ func doDeleteObject(bs *bucketService, keys simplecrypto.Keys, filename string) 
 	encryptedFilename := encToDecPaths[filename]
 
 	if encryptedFilename == "" {
-		return fmt.Errorf("File: %s not found in bucket", filename)
+		return fmt.Errorf("file: %s not found in bucket", filename)
 	}
 
 	return bs.deleteObject(encryptedFilename)
